Make the cscope symbol search better
Some times cscope does not find the definitions of symbols, but it find
their use cases which might include the actual definitions of the symbols.
Thus, add the -L0 check in cscope to account for these cases.

Also account for cases where cscope might get confused with callbacks that
have the same name as the target symbols. For example, for the "open"
symbol, cscope will return all the cases a callback with the same name is
used. E.g., .open = <callback>.